Use DuckDB SQL for access policy expressions (#3006)
* Use DuckDB SQL for access policy expressions

* Better type check

* Update docs

* Fix test
